Added fields to Django Admin to be able to edit a User's Organization and Project membership in the User admin.
The ManyToMany relationship between Organizations/Projects and Users
is defined in those models (and not the User model).  This makes sense as
the User model is defined by Django.  By default, the ManyToMany field
is displayed in the model admin for the model where it is defined.  I
had to do a little bit of hackery to allow for editing the relations
through the User admin as well.

http://blog.abiss.gr/mgogoulos/entry/many_to_many_relationships_and
was very helpful in guiding me how to do this.