Changed Story admin to use filter_horizontal for defining Story to Asset relations.
This widget is better than the inline we were using and the inline isn't
neccessary anymore since we're not using the through class.

The through class was used to provide a weight to order assets within a
story, but this is now done in the section.  Assets are associated at
the story level just to build a bucket of assets for potential use
in a story.