Cleaned up tests.
Moved tests dealing with SectionResource to SectionResourceTest. Even
though they're dispatched from StoryResource, most of the code lives
in SectionResource, so it made sense to put the tests with other tests
for that class.

Implemented SectionLayoutModelTest.test_get_template_contents as I forgot
to do this during the last commit.