Management command to copy Handlebars templates to a place where jasmine tests can find them
This isn't awesome, but it's the most DRY way I could think of to make
the Handlebars templates available both to the Django view's template and
the Jasmine tests via the jasmine-jquery plugin.

Long term, it might be better for SpecRunner.html to get rendered through
a Django template, but this seemed like a good quick and dirty solution.