Modified section rendering to use new-style SectionAsset relationships
For the first version of the viewer, the assets were just displayed in order
by weight.  Now, the assets are mapped to containers within a section
layout.  I made a special {% container %} template tag to display the
rendered asset or a placeholder div if context isn't provided.

This commit also includes cleaning up a lot of indentation weirdness that
I found.