Switched from using signals to just calling hook methods.
This is more consistent with how Tastypie works in general. With the signals
there wasn't a clean way to bind them to specific subclasses, so the
handlers had to do a lot of type checking.

I think this is better and in the worst case scenario I might have to
call to specific superclass implementations in the subclass' hook
implementations.