Let the router trigger all the rendering.
This eliminates an extra render for the main view and for passing the
initial workflow step from the Django view.

As part of implementing this, I realized that the best way to handle the
internationalized URLs would be to detected them on page load and set
the routers root accordingly.  This greatly simplified the route matching
code.