Moved and renamed view mixins.
Since 'Mixin' is in the view name, and there aren't very many, having
a separate namespace for them just added length and not a lot of utility.

RichTextEditorMixin seemed to fit better in views.js since it operates on
the builder views and not on Backbone Forms form or form element views.