Encapsulate the spacer div in the section view.
This makes cleaning up the spacers a lot easier, because the spacers
disappear when the thumbnail goes away when a section is deleted.
The other thumbnails can listen for a signal and re-render to set the
index data on the spacer elements.

In order to get this working, I had to do better cleanup of my views
event bindings because the re-render event was getting called for
thumbnail views for sections that no longer exist.

http://lostechies.com/derickbailey/2011/09/15/zombies-run-managing-page-transitions-in-backbone-apps/
was really helpful for providing patterns for cleaning up the zombie
views.