Edit asset captions.
In order to do this I needed to handle posting an image property to
AssetResource in both a data_uri and a relative URL.  The latter is just
posting back the value returned by a get request.  It's ignored but has to
be ignored explicitly because Backbone always sends a full set of attributes.