Patial fix for #151
The callback for Jeditable now saves the model in a way that supresses
a bubbling up of the sync signal.  This eliminates unneccessary view
creation and rendering.

Also, associate newly created assets with their container via an
attribute (that gets ignored by the API).  For assets that are fetched
from the server, the SectionAssets collection's parse method creates
this attribute.