Disable superuser creation prompt on syncdb
I had originally tried to work around the issues with the post_save
signal handler that creates a new UserProfile when a user is created,
but it seems like attaching the superuser creation signal handler to
post_migrate causes this prompt to be shown when running tests.

For now, just disable it.