Keep the weight attribute of SectionAsset.
This is neccessary to preserve backward compatibility to support stories
created on staging before we had the notion of containers and layouts.

This required editing a past South migration which might be considered
poor form, but I can't really think of a good way to do this otherwise.

Since the app isn't deployed anywhere else I think it's ok.