Render section assets not associated with a container
Closes #143

This implements a strategy for allowing sections to have assets not
associated with a particular section container.  This is needed to support
displaying stories created in the Django admin before the builder and
its notion of layouts and containers was implemented.

The rendering logic is largely moved out of Section.render_html and into
the container template tag