Story template loading in existing instance
Allow the stock story templates to be loaded from fixtures in an instance
that already has content (i.e. the current staging instance).

The stucture for a story template is just provided by a story itself.
This creates a problem with loading the templates into an existing instance
because of having to conflicts on the primary keys of the template story
models.

I worked around this by using Django's natural key serialization (see
https://docs.djangoproject.com/en/dev/topics/serialization/#natural-keys)
to refer to related models by their UUID key instead of the primary key.
This is a bit of an abuse of the functionality, since the UUID isn't a very
natural key.

There was a bug in Django's dumpdata management command that didn't allow
serializing self-referencing models (in our case Section) when using
natural keys (see https://code.djangoproject.com/ticket/16317) that has
since been fixed.  I backported the patch to Django 1.3 and make it
applicable using the apply_django_patches script.

I dumped new versions of the fixture JSON for section layouts and story
templates.  I had to manually edit files to merge in the models from
storybase_help into the ones from storybase_story, to break out the
layout and container models and finally to replace references to primary
keys with "pk": null