Cleaned up layouts of section editing in builder
A big part of this is centering the section assets in a 940px column,
then calculating the height of the asset containers based on common aspect
ratios of screens/images.

To make it easier to change things up, I've started to employ the LESS
dynamic stylesheet tool. The builder template uses a pattern that
compiles the LESS styles client-side in debug mode, and expects them
precompiled in production. I wrote a simple shell script to automate the
compilation of the LESS styles in scripts/buildless, but should probably
eventually move to something like django-compressor or django-pipeline.