Use a separate path for the connected builder
Instead of just passing the template in the querystring, have a
separate URL that causes the template to get passed as a keyword
argument to the view.  This eliminates a bit of template logic and
makes the URLs a lot cleaner.

Implementing this involved reworking the getRoot() function. I went ahead
and generalized this so it seemed to make sense to move it to the utils
namespace.