Lookup section thumbnail views by section cid
Closes #210

Section thumbnails weren't being stored in this._thumbnailViews when
the section was copied from the template.  That's because
this._thumbnailViews was being keyed by section.id, which isn't set
for the template sections because they weren't yet saved.  This fix
uses section.cid as the key for this._thumbnailViews when the section has
not yet been saved.