Re-delegate events on render
Closes #228

The story builder view's element is removed from the DOM when you switch
to any of the other workflow steps.  This causes the event bindings to be
lost.  So, we need to explicitly call delegateEvents().  In the long run
it might be better to just hide the view elements instead of removing them
from the DOM with the call to empty() in AppView.