Make Backbone routing work without pushState
Partial fix for #235

IE9 (and likely some other browsers) don't support the pushState API,
so we need to make routing work both for real paths and hash-based paths.

In order to do this, we need to make the story ID work both as part of the
root and as part of the path.  This required defining the routes dynamically
in the router's initialize method instead of declaring them as a property.

As part of this commit, I backported the replacement of
storybase.utils.getValue with _.result from other branches to ease merging.