Make help popup display via a hover event.
In order to do this I had to stop using the SimpleModal plugin because
when it opened the popup, it was doing something to the mouse so the
hover event kept firing.

I wrote quick and dirty event handler to display the popup, but it should
probably be generalized as these need to be shown in other places as well.