Use declarative syntax to override auth form
Closes #279.

Originally, I was altering the max_length property in the constructor
of my AuthenticationForm subclass.  This had two problems.  First, setting
max_length doesn't set the maxlength attribute of the HTML element
generated by the widget.  More importantly, calling the parent class'
constructor causes the char field to be constructed which gets a length
validator attached to it at the original field maximum length of 30.
Using the declarative syntax to override the form field properties
avoids both of these problems.