A timezone-aware version of DirtyFieldsMixin
This was needed when experimenting with mixing DirtyFields into the
Story class.  I think the API gets timezone-aware datetimes and
DirtyFieldsMixin.get_dirty_fields was failing when trying to compare
timezone-aware datetimes with timezone-naive ones.  I modified
DirtyFieldsMixin to use an overridable comparison function that is
overrided in TzDirtyFieldsMixin to handle this case.