Update the flow of publication step
Each "step" (accepting the legal agreement, selecting a license, and
selecting a featured image) is now shown and supports an expanded or
collapsed mode.  There is a message at the top that is updated with the
remaining steps that need to be taken.

I cleaned up and broke out the logic for rendering/toggling the
different elements in the parent view.