Cleanup translation string marking
This cleans up some of the updates to builder copy that @jwirfs-brock made.

Generally, use of the {{gettext}} Handelbars tag should be placed within
HTML tags, e.g. <p>{{gettext message="Blah blah blah"}}</p>. So the
translator doesn't have to deal with HTML in their strings to translate.

In some cases, like making text strong, it doesn't make sense to break up
the string to be translated into pieces. In this case, the translator will
just have to be observant of a little HTML in their translation strings.
We'll also use {{blockgettext}} to mark this up in the Handlebars templates.

<p>
{{#blockgettext}}This contains <strong>some HTML</strong>{{/blockgettext}}
</p>