Basic drawer functionality.
This is the basic interactions for the drawer.  It lives in DrawerView.
Other views that will live in the drawer need to be registered with
DrawerView.registerView and implement a drawerButton, drawerOpenEvents and
drawerCloseEvents method.

The general concept of the drawer is that it tries to stay as agnostic to
the content of the inside views as possible.  DrawerView fires
``open:drawer`` and ``close:drawer`` events when the drawer is opened or
closed.  AppView listens for these events to set a class on the main
container view which lets us use CSS for setting the width of the drawer.