Integrate grid into style.css
This is done by changing grid.css to be a LESS file (really this just
involved changing the filename extension) and importing it at the top
of base.less.

The real motivation for this is to let us use the different grid classes
as mixins. This lets us use just semantic class names instead of peppering
our markup with .container_12 and .grid_6.  Furthermore it lets us have
the ease of using the grid convenince classes without having to put the
presentational class names in the Backbone view declarations.  Ultimately,
it should make the code more maintainable because all the presentational
stuff lives in the same place (the LESS stylesheets) and there should be
minimal editing of the HTML templates or Backbone views to change the
grid.

The only drawback is that using the grid classes as mixins replicates the
style definitions, which would increase the size of the final CSS.  I'm
thinking this trade-off is worth it in terms of making the styles more
maintainable.

This commit also includes giving most of the builder views a common CSS
class name of 'view-container' (which gets the .container_12 mixin in
builder.less).