s/defaults/options/
Backbone's View._configure function extends existing this.options on the
view object rather than clobbering them, providing a provision for
default options declared on the view.  This makes the defaults property and
the subsequent call to _.defaults(this.options, this.defaults)
unnecessary.  Instead, I just declare an options property of the view with
the default options.