Clone commonOptions
The view constructor replaces this.options with options and extends options
with the properties of the options declared on the View.  So passing
commonOptions to subsequent view constructors gives each subsequent view
the options modified by the previous view.  Cloning the commonOptions object
before passsing it to each view constructor avoids this problem.

Closes #360.