Give min-height to #story-list
This fixes the bug where #app wasn't given the correct padding-top.

On an initial page load, the story sections are loaded via AJAX and
the section list doesn't have a height until the sections are loaded
and the section list view is re-rendered.  The padding of the #app wrapper
is calculated before the section list gets populated so the height
calculation for the header is wrong.  Giving the section list a min-height
equal to it's final rendered height lets the dynamic padding happen before
the section list has content.