Updated drawer toggling
The tabs now function to both open and close the drawer as well as toggle
between views in the drawer if the drawer is already opened.

This commit required reworking the API for registering/unregistering views
with the drawer.  I also added some simple documentation and tests for the
drawer.