Styling for assets in drawer
Style the assets in the unused assets drawer.  In this commit I made the
thumbnail images a bit wider, added default icons for when no thumbnail is
available, added borders to the elements and centered everything in the
container.

I also had to work around an issue where jQuery UI's draggable elements were
getting hidden when their parent container has overflow set to hidden or
scroll.