Add slug field to SectionLayout model
This is needed to add styles for the fancy layout selection dropdown. A
human-readable class name is way better than a silly UUID class name.

Eventually, we can do away with the section_id field and use the slug field
instead.  I'll be making it unique in the next couple of commits.