Hide search input
Search isn't yet wired up (we decided that the explore view provided
much of the needed functionality).  Hide the search input for now.

In the interest of expedience and not having to rework this once search
is enabled, I just put in a div with the same height as the search form.
Blech.

Closes #317