Templatized asset help templates.
The type-specific help for assets (which is appended to any template-related
help and shown when the user clicks the help icon) seemed like something
that might get edited a couple of times, so it made sense to break it out
from the source code.  In the future we could wire this up to the Help model
in Django if we want to make these editable via the Django admin.

I also cleaned up the logic for selecting the type.  Previously, it was
only looking at the suggested type option for the view, but this option is
used to set the bound model's type attribute in the view's initialize()
function, so we can just check the model's type attribute.  Also, doing
it this way lets us show type-specific help messages in cases where the
user can override the suggested type.

Closes #225 as one of the templates includes a message to use the video's
URL and not the embed code.