Wire in featured story, fix recent list
In previous commits, I had misunderstood how featured and recent stories are
identified. ``Project.curated_stories`` is a list of stories that will show up
in the featured list.  Currently this is just limited to a single story by
``FeaturedMixin.featured``.  Recent stories should be populated from
``Project.stories`` which is the reverse relationship from ``Story.projects``.
The list is sorted and limited by ``RecentMixin.recent_list``.

Partial implementation of #394 and #395.