Replace "Citizen Atlas" with "Floodlight"
In the Open Graph Meta Tags, "Citizen Atlas" was now hard-coded.  The
tags now use the value of ``settings.STORYBASE_SITE_NAME``.  The
description text is also now a setting,
``settings.STORYBASE_SITE_DESCRIPTION``.  Finally, I added a context
preprocessor to add the site domain and protocol to the RequestContext.

Closes #445.