First pass at IFRAME uploads
Internet Explorer <= 9 doesn't support the HTML5 File API, so we have to
append a hidden IFRAME and submit the form to the IFRAME to fake an AJAX
upload.

I've only added support for this for asset uploading.  Still to do is
implementing this for the featured asset and dataset uploads.

Partial fix for #236