Don't retrieve default featured asset, set it
When a story doesn't have a featured asset explicitly defined, we've
been using the first image asset.  Previously, the code did this each
time the featured asset was requested.  It's kind of expenseive in terms
of database use, so it seemed better to do it once when the story is
published instead of each time the story data is fetched.  I added a
pre_save signal handler to auto-assign a featured asset if none is
explicitly specified.

The pre_save signal should be triggered for all existing stories, just
to make sure a featured asset is set (since it will no longer have a
default retrieved on access).  This can be done by just saving each
published story:

    from storybase_story.models import Story
    for s in Story.objects.filter(status='published'):
        s.save()

Partial fix for #455