Fix the layout of the publish step
Classes that controlled layout were still using class names from the
1140 grid system instead of our custom grid.

I also made the element in which the subviews are appended an option
instead of just a hard-coded ``.left``.

Finally, I added a header to the sharing sidebar to git it some space
at the top and gave the input with the sharing URL a bottom margin to give
it a little breathing room from the AddThis widgets.

Closes #470.