Remove bullets, align buttons
In the asset list displayed allowing the user to select one of the assets
in the story as the featured asset, a bullet was shown next to each asset
due to the default list styles.

I removed the bullet and also aligned the "Set as featued image" button to
the images.

Closes #471.