Refactor dehydrated versions of M2M fields
This commit moves much of the heavy lifting to the Story model, mainly
to keep all the knowledge of the caching strategy in one place. It's
also way more DRY.  Finally, I started storing the language code as part
of the cache key because the values will differ depending on the
language.