Embedable widget for stories
This implements the back-end code, and embedding Javascript for a story
widget.  A story can now be embedded with code like this:

        <a href="http://floodlightproject.org/stories/asdas-3/" class="storybase-story-embed" data-height="450">asdasa on Floodlight</a>
        <script src="http://floodlightproject.org/static/js/widgets.min.js" type="text/javascript"></script>

This code loads a Javascript (widgets.min.js) that dynamically hides the
placeholder link and adds an iframe of the widget HTML.  The widget HTML
is at the path /stories/STORY_SLUG_OR_ID/widget/ and is generated by
``storybase_story.views.StoryWidgetView``.  This commit also adds a
``story_embed`` template tag in the ``story`` tag library that will
generate the embed code like the one above.  The HTML is rendered by the
``storybase_story/story_embed.html`` template.

The widget view is styled by ``widget.less``/``widget.css`` which
contains a lot of code similar to the styles for the tile view of the
story explorer. At some point, it might make sense to refactor the LESS
styles to not have repeated styles, but at this point, I'm keeping them
separate for clarity and expedience, especially since the style of the
embedded widget seems something likely to change.

Note that when changing ``widget.less``, it needs to be compiled with
lessc to generate an updated ``widget.css``.  ``widgets.js`` needs to
be minified (I'm using Google Closure Compiler) to generate
``widgets.min.js``

Implements the back-end code for #459 (user story 40)