Refactor addthis widget rendering into a tag
Making the AddThis widget a template tag instead of an included template
removes some ugly logic in the templates that was needed to prepare the
context for the included template.  It also allows for one tag that
takes arguments for use in different contexts (different widget sizes,
including the JavaScript or not) different rather than having separate
templates for each (or having a bunch of logic in the template).

The big slow-down in doing this refactor was having to write a custom
template tag that supported keyword arguments.  If we were using Django
>=1.4, this wouldn't be hard since the inclusion tag helper supports
keyword arguments.

I added some utility functions in ``storybase.templatetags`` to help
with parsing keyword arguments in custom template tags.  They're
largely taken from http://djangosnippets.org/snippets/1113/.

This provides functionality for #459 and #475.