Add endpoint for sharing widget
This endpoint renders the contents of a story sharing widget, trying to
offer a DRY way to provide this content to both server-side Django and
client-side Backbone views.

This currently lives at /stories/SLUG_OR_ID/share-widget/.

This returns an HTML fragment rather than a full HTML page (though it might
make sense to conditionally wrap the output in an page HTML for other
use cases).  It is designed to be fetched via an asynchronous JavaScript
request.

Note that the AddThis JavaScript script tag is not included in the returned
code as it is assumed to be loaded and executed by the client JavaScript.