Add translations and asset text to index
Stories need to be discoverable by their content and in all languages.

This adds the title and summary in all languages, the section titles
in all languages and the asset text.  I added a ``strings`` method to
the ``Asset`` model and the ``AssetTranslation`` model to help with
this.  For now, I'm only worrying about the ``HtmlAsset`` model since
this is how most text is stored.