Search index updates
Update StoryIndex when a story's related models change and provide spelling
suggestions for search.

This commit adds signal handlers to update the StoryIndex when the text of
an asset changes, when the text assets that are part of a story change and
when the tags associated with a story changes.

Support for spelling suggestions for search results has also been added.
I followed the instructions in the Haystack documentation to wire this up
(see http://django-haystack.readthedocs.org/en/latest/installing_search_engines.html#spelling-suggestions), but using a
``FacetCharField`` for the index field that would be used to generate the
spelling suggestion index/dictionary didn't work.  I had to create a
field type following the example at http://wiki.apache.org/solr/SpellCheckingAnalysis.
As a result, I had to update the template used to generate schema.xml
(``search_configuration/solr.xml``) to include the XML for the field
definition.  I also had to update the custom search backend,
``storybase_geo.search.backends.Solr2155SearchBackend``, to include
logic for correctly setting the field type to the custom field when using
the ``build_solr_schema`` management command.

For a reason I don't quite understand, the default ``SearchView`` in
Haystack only shows spelling suggestions if there are search results.  This
is the opposite of what we want, which is to show spelling suggestions if
there are no search results.  So, I had to make a custom SearchView
subclass, ``storybase.search.views.StorybaseSearchView`` that implements
the desired behavior of providing spelling suggestions.

Implements part of #464.