Search result refinements
Implement some of @bpawlak's suggestions including:

* Show "Tags" label for list of story's topics and tags.  I'm a little
ambivalent about this because in the tagging step of the builder we make
the distinction between "topics", the canonical list of categories in the
site, used in the story explorer, and "tags", a mechanism for users to
insert keywords into the search index.
* Don't show a title at the top of the page when a search query has been
performed.  Having a title that simply says "Search" or "Search Results"
just takes up space when a query has been performed because a user obviously
knows that this is a search page. I've preserved the title for when the
user visits the page directly.
* Show search query in the upper-right-hand search box. I didn't go as far
as @bpawlak's suggestion to remove the search form above the search results
because I feel strongly that users expect the search form here based on
typical search implementations.  I also feel like if users want to modify
their search results, they'll shouldn't have to go all the way up to the
upper right hand corner.  However, it would be very easy to hide the search
form if we *really* want to do that.

Partial implementation of #464