Refactor custom id field retrieval
Wrap the logic for retrieving a model's custom ID field (the uuid fields
I'm using to hide the internal pk and to make objects portable via fixtures
or potentially across instances) in a method so that subclassses of this
view can provide different logic for retrieving the ID field name.

This is because, in the vast majority of cases, the custom id field is
just the model name lowercased with "_id" appended.  However, this isn't
always the case (specifically ``UserProfile.profile_id``.