User profile view
This implements a user profile view that lists stories authored by a user.

In order to implement this, I refactored a number of model mixins in the
``storybase_user`` app and the ``storybase_user/detail_base.html``
template to be able to use this template for projects, organizations and
user profiles.

I also updated ``ModelIdDetailView`` to be able to lookup models by a
shortened uuid field using the ``shortuuid`` python package.

Implements #461, user story number 98