Refactor Asset image rendering
I pulled the methods used to render thumbnail images from Asset and put
them into ImageRenderingMixin.

I pulled the default image retrieving methods from FeaturedAssetsMixin and
put them in DefaultImageMixin

The reason for this is to be able to use these implementations and
interfaces for classes that aren't Assets but behave like assets like a
user profile image.