Adjust margin and width of save button
The #last-saved element had an explicit width set in builder.less. This
was required to make the last saved timestamp text wrap to fit a space
approximately the same size as the button. Unfortunately, the
combination of the 100px width and the 10px right margin made this
element get pushed below the other workflow bar elements. This caused
the button to get hidden altogether because of the higher z-index of the
section list bar.

I reduced the right margin of the #last-saved element and also made the
width a max-width value.  I also added an explicit style to supress
the italic font-style that was being inherited in IE.

Closes #491