Parse API's ISO-8601 date strings in IE
The Backbone app gets the date from the server in a string like
"2012-12-03T12:35:04.088137-06:00".  IE fails when constructing a new
Date object from this string.

This is because IE doesn't support parsing non-strict ISO-8601 times
(i.e. ones that include milli/microseconds).  datetime.isoformat, which
is used by Tastypie to serialize DateTime fields in ISO-8601 format,
includes micrseconds as part of the string (See
http://docs.python.org/dev/library/datetime.html#datetime.datetime.isoformat)
See https://github.com/toastdriven/django-tastypie/pull/460 for a
proposal to address this in Tastypie.

I decided to address this client-side by checking if the date parsing
fails and then using a regular expression to remove the microsecond
portion of the time in the string.

As part of this implentation, I refactored the date parsing logic into its
own method, LastSavedView.parseDate, which makes it easier to refine the date
parsing logic if needed and also simplifies the use of the tenery
operator in the assignment of this.lastSaved in LastSavedView.initialize

Closes #237