Added some helper filters
I added some helper filters to the storybase_tags module.

classname outputs the class name of an object, in most cases a model.
camelsplit splits a camel-case string at the transition between
lowercase and uppercase letters. These are meant to be used together to
help make polymorphic templates where strings in the rendered template
come from class names.  So, if you had a model with the class name
UserProfile, you could print "user" in the rendered template using a
chain of filters like this:

    Share this {{ object|classname|camelsplit|first }}

These tags help with the implementation of #499