Nudge tour popups in bound event
Previously, I had hacked guiders-1.2.8.js to move over a guider element
if it's offset was to the left or right of the window.  This change
moves this nudging to an event handler bound with jQuery's ``.bind()``
to the guider's element.  We can't do this in the the guider's
``onShow`` callback because this is called before the guider element is
positioned in the DOM.  Instead, we bind the function that does the
nudging in the ``onShow`` callback to the ``guiders.show`` event that is
triggered at the end of the ``guiders.show`` function.

This also addresses #415 because the element position change was working
in an indeterminate way, due to some kind of timing issue, making the
guider element render with portions outside of the window. Waiting until
the jQuery event is triggered seems to correct this issue.

I refactored the logic to determine whether or not to show the tour into
a method to get rid of a long and hard to read assignment statement with
multiple ternary operators.

I also added an optional ``STORYBASE_FORCE_BUILDER_TOUR`` setting that
can be set in the Django settings module to force always showing the
tour.  This is just for debugging purposes.  I thought about making it a
query string parameter, but this breaks the Backbone routing.