Workaround for #415 in IE
Apparently e3021f8496a5178300cb64b6de6a5fb53aeb14d2 didn't fix this
issue all the time.  This is a horrible hack that delays and then
readjusts the position of the first guider element.  In my testing, I
haven't seen the issue for any of the guders except the first one.