Form class for translated models
This is a first pass at a form class that handles the translated fields
as if they were all on the same model instead of spread out between two
models.

I added some helper methods to the TranslatedModel class to make it easier
to implement this.

CreateOrganizationForm is an example of using TranslatedModelForm.

This functionality is needed as part of #458 and #463, but when fully
implemented and tested could help refactor a lot of the translation code.