Publish existing projects and organizations
After schema migration 0017_add_publication_status_and_contact_info, the
newly added status field is set to the default of 'draft'.  Since we
previously didn't have any publication status control on these models, we
assume that we want them to be published.  This data migration sets the
status of all Organizations and Projects to 'published'.