Don't show "Publish" button
Don't show the "Publish" button for a story that has never been published.

In order to detect this case, a never_published property was added to the
PublishedModel class.

This commit also adds some unit tests for the AccountStoriesView to check
the logic for displaying different publication changing buttons.

Closes #460, which was re-opened because of the loophole that allowed users
to bypasss the publication flow in the builder by clicking the "Publish"
button in the stories listing.  See this URL for details:

https://github.com/PitonFoundation/atlas/issues/460#issuecomment-11342637