Add a message explaining profile image setting
Add a message on the user profile page explaining how the profile image is
set, shown only for logged-in users when viewing their own profile page.

In the future, it might be nice to style this message so it looks different
than normal content text. I've created an issue for the general need for
alert styles as #543.

Addresses #461