Show link to public profile link
As per @bpawlak's suggestion at https://github.com/PitonFoundation/atlas/issues/461#issuecomment-11453441,
show the public profile link at the top of the account information page.

This also seemed like a better place to display the note about how the
profile image is selected, so I moved that message out of the public profile
template and into the template for the account summary.

Addresses #461