Show embed widget height attribute in default code
By specifying a ``data-height`` attribute on the link in a story's embed
code, the user can control the height of the IFRAME element that is
ultimately rendered by the widget JavaScript. This was a previously
undocumented feature. The ultimate way of helping users control the height
of their widget will be to create a view that lets the user fill out a
form that will change different properties of the embed code and to
preview what the widget will look like with those attributes set.  For now,
however, showing the default value of the ``data-height`` attribute will
give early adopters a hint that they can adjust this value to change the
height of the embedded widget.

Addresses #529