Allow re-showing the tour
This wires up the "View the tour again" button.

In order to make this work, I had to add an optional callback argument to
the select:workflowstep event and the AppView.updateStep method in order to
trigger showing the tour only after the builder had switched back to the
correct workflow step.

I also had to add a 'do:show:tour' event to trigger showing the tour.

Finally, I added a force argument to the various tour showing methods to
override the default logic checks for showing the tour.

Addresses #405