Make editor dialog buttons buttons
I had just copied the markup from the example, but it's more semantically
correct to make the "OK" and "Cancel" buttons BUTTON elements instead of A
elements.  Doing this also styled them consistently with other buttons
in the builder, though I had to adjust the styling a little bit to fit
in the smaller space.

Addresses #231.