Replace the "View tour again" button with an icon
Replacing the "View tour again" button with an icon makes it compete less with the other help content.  I added a title attribute that will cause "View tour again" to be displayed by the browser on mouseover.

Users might still miss it, so I thought it made sense to show how to view the tour again as part of the tour.

In order to implement the styling of the button, I added a button.reset style (based on the button-reset mixin) that removes the default button styling.  I also layed some groundwork for mixins that match the Font Awesome icon classes.

Finally, I added styles to show the right-facing tour popup arrows, which were missing from the first pass at guider styling.