Only show the view tour again button build step
To be forward thinking about this, I conceptualized "help actions" which
could be any kind of buttons that interact with the larger system that are
shown at the top of the content inside the help drawer.

I created a simple API for views in the drawer that have actions associated
with them.  This includes a do:set:helpactions signal that sets a view that
renders and defines UI events for the help actions and a do:clear:
helpactions that removes the view.

I added a HelpView.setActions and HelpView.clearActions method to act as a
handler for the events and a HelpView.renderActions
to break out the call to render the actions view.  This is because I call
setElement to set the element before rendering and in some cases the
this.actionsView isn't set, so the rendering logic is a little more
involved than just calling this.actionsView.render()

Finally, since the view tour again button is no longer shown in steps other
than the build step, I removed the logic that switched to the build step
before showing the tour.

Closes #508